compressing_microservice/requirements.txt
fastapi
pillow
moviepy 
ffmpeg-python
celery[redis]
redis
importlib-metadata==4.13.0
websockets